feat: add cie_timestamp, cie_encrypt, cie_decrypt public API
- cie_timestamp: standalone RFC 3161 timestamp (SHA-256 + TSA, no card needed)
- cie_encrypt: RSA-OAEP / hybrid AES-256-GCM+RSA-OAEP using card public key
- cie_decrypt: RSA-OAEP / hybrid AES-256-GCM+RSA-OAEP using card private key
  via PC/SC IAS raw RSA operation + software OAEP unpadding

Exports added to linker/libopencie-pkcs11.map and .exp.
meson.build: new CSP sources added under build_sign_sdk guard;
pkcs11_inc now includes sign-sdk/include and sign-sdk/src when
build_sign_sdk is true so tsa_client.h and its ASN.1 deps resolve.

+++ pkcs11/src/sign/cie_verify.cpp
@@ -30,7 +30,8 @@ long CIEVerify::verify(const char* input_file, VERIFY_RESULT* verifyResult,
     }
 
     ret = cie_sign_verify_set(ctx, CIE_SIGN_OPT_INPUTFILE_TYPE,
-                              reinterpret_cast<void*>(CIE_SIGN_FILETYPE_AUTO));
+                              reinterpret_cast<void*>(static_cast<uintptr_t>(
+                                  CIE_SIGN_FILETYPE_AUTO)));
     if (ret != 0) {
       throw ret;
     }
@@ -111,7 +112,8 @@ long CIEVerify::get_file_from_p7m(const char* input_file,
     }
 
     ret = cie_sign_verify_set(ctx, CIE_SIGN_OPT_INPUTFILE_TYPE,
-                              reinterpret_cast<void*>(CIE_SIGN_FILETYPE_AUTO));
+                              reinterpret_cast<void*>(static_cast<uintptr_t>(
+                                  CIE_SIGN_FILETYPE_AUTO)));
     if (ret != 0) {
       throw ret;
     }
